Usage of Factory class.  UserProfile now calls with EXTRA productId to ItemPageActivity (not sure what the name is).  Updated the productId from int to long  Fixed some ui UI issues

+++ app/src/main/java/com/benefit/services/ProductService.java
@@ -40,11 +40,11 @@ public class ProductService extends ViewModel {
         this.productsCollection.add(product);
     }
 
-    public LiveData<Boolean> deleteProduct(int productId) {
+    public LiveData<Boolean> deleteProduct(long productId) {
         return this.databaseDriver.deleteDocumentsByField(COLLECTION_NAME, "id", productId);
     }
 
-    public LiveData<Product> getProductById(int productId) {
+    public LiveData<Product> getProductById(long productId) {
         return this.databaseDriver.getSingleDocumentByField(COLLECTION_NAME, "id", productId, Product.class);
     }
